backend/requirements.txt
annotated-types==0.7.0
greenlet==3.2.3
pydantic==2.11.7
pydantic_core==2.33.2
python-dotenv==1.1.1
SQLAlchemy==2.0.42
typing-inspection==0.4.1
typing_extensions==4.14.1
